# WMS_WebAPI/WMS_WebAPI/Controllers/DockMasterController.cs
using System;
using System.Data;
using System.Data.SqlClient;
using System.Linq;
using System.Web.Http;
using WMS_WebAPI.Models;
using WMS_WebAPI.Models.Context;

namespace WMS_WebAPI.Controllers
{
    public class DockMasterController : ApiController
    {
        WMS_Entities _context = new WMS_Entities();
        CommanListToDataTableConverter ConvertDataTable = new CommanListToDataTableConverter();
        string connectionstring = System.Configuration.ConfigurationManager.ConnectionStrings["ConStr"].ConnectionString;
        [HttpGet]
        [Route("api/DockMaster/Dock_Select")]
        public IHttpActionResult Dock_Select()
        {
            try
            {
                var data = _context.Dock_Select().ToList();
                if (data == null)
                {
                    return NotFound();
                }
                return Ok(data);
            }
            catch (System.Exception)
            {

                return BadRequest();
            }

        }

        [HttpPost]
        [Route("api/DockMaster/Dock_insert")]
        public IHttpActionResult Dock_insert(cls_DockMaster obj)
        {
            try
            {
                var data = _context.Dock_insert(obj.dockID,obj.dockName,obj.dockNo,obj.isActive,obj.wareHouseID).ToList();
                if (data == null)
                {
                    return NotFound();
                }
                return Ok(data);
            }
            catch (System.Exception)
            {

                return BadRequest();
            }

        }

    }
}