# WMS_WebAPI/WMS_WebAPI/swagger/swaggerjson - Copy.json
{
  "swagger": "2.0",
  "info": {
    "version": "v1"
  },
  "host": "localhost:50191",
  "schemes": [ "http", "https" ],
  "paths": {
...
    "/api/CustomerContacts/GetCustomerContactsList": {
      "get": {
        "tags": [ "CustomerContacts" ],
        "operationId": "CustomerContacts_GetCustomerContactsList",
        "consumes": [],
        "produces": [ "application/json", "text/json" ],
        "parameters": [
          {
            "name": "CustomerID",
            "in": "query",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": { "type": "object" }
          }
        }
      }
    },
    "/api/CustomerContacts/CustomerContractWarehouse_List": {
      "post": {
        "tags": [ "CustomerContacts" ],
        "operationId": "CustomerContacts_CustomerContractWarehouse_List",
        "consumes": [ "application/json", "text/json", "application/x-www-form-urlencoded" ],
        "produces": [ "application/json", "text/json" ],
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/cls_CustomerContractWarehouse" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": { "type": "object" }
          }
        }
      }
    },
    "/api/CustomerContacts/CustomerRateList_Insert": {
      "post": {
        "tags": [ "CustomerContacts" ],
        "operationId": "CustomerContacts_CustomerRateList_Insert",
        "consumes": [ "application/json", "text/json", "application/x-www-form-urlencoded" ],
        "produces": [ "application/json", "text/json" ],
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/cls_customerRateList" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": { "type": "object" }
          }
        }
      }
    },
    "/api/CustomerContacts/CustomerStorageArea_Select": {
      "post": {
        "tags": [ "CustomerContacts" ],
        "operationId": "CustomerContacts_CustomerStorageArea_Select",
        "consumes": [ "application/json", "text/json", "application/x-www-form-urlencoded" ],
        "produces": [ "application/json", "text/json" ],
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/cls_customerRateList" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": { "type": "object" }
          }
        }
      }
    },
    "/api/CustomerContacts/CustomerRateList_Service": {
      "post": {
        "tags": [ "CustomerContacts" ],
        "operationId": "CustomerContacts_CustomerRateList_Service",
        "consumes": [ "application/json", "text/json", "application/x-www-form-urlencoded" ],
        "produces": [ "application/json", "text/json" ],
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/cls_customerRateList" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": { "type": "object" }
          }
        }
      }
    },
...
    "/api/CustomerLimit/customerlimit_insert": {
      "post": {
        "tags": [ "CustomerLimit" ],
        "operationId": "CustomerLimit_customerlimit_insert",
        "consumes": [ "application/json", "text/json", "application/x-www-form-urlencoded" ],
        "produces": [ "application/json", "text/json" ],
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/cls_CustomerLimit" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": { "type": "object" }
          }
        }
      }
    },
...
    "/api/CustomersStock/CustomerView_GetData": {
      "post": {
        "tags": [ "CustomersStock" ],
        "operationId": "CustomersStock_CustomerView_GetData",
        "consumes": [ "application/json", "text/json", "application/x-www-form-urlencoded" ],
        "produces": [ "application/json", "text/json" ],
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/cls_CustomersStock" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": { "type": "object" }
          }
        }
      }
    },
...
    "/api/DockMaster/Dock_insert": {
      "post": {
        "tags": [ "DockMaster" ],
        "operationId": "DockMaster_Dock_insert",
        "consumes": [ "application/json", "text/json", "application/x-www-form-urlencoded" ],
        "produces": [ "application/json", "text/json" ],
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/cls_DockMaster" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": { "type": "object" }
          }
        }
      }
    },
...
    "/api/InwardUpdate/InwardUpdate_insert": {
      "post": {
        "tags": [ "InwardUpdate" ],
        "operationId": "InwardUpdate_InwardUpdate_insert",
        "consumes": [ "application/json", "text/json", "application/x-www-form-urlencoded" ],
        "produces": [ "application/json", "text/json" ],
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/cls_InwardUpdate" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": { "type": "object" }
          }
        }
      }
    },
...
    "/api/Login/SP_Login": {
      "post": {
        "tags": [ "Login" ],
        "operationId": "Login_SP_Login",
        "consumes": [ "application/json", "text/json", "application/x-www-form-urlencoded" ],
        "produces": [ "application/json", "text/json" ],
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/cls_UserLogin" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": { "type": "object" }
          }
        }
      }
    },
    "/api/LotSearch/LotSearch_List": {
      "post": {
        "tags": [ "LotSearch" ],
        "operationId": "LotSearch_LotSearch_List",
        "consumes": [ "application/json", "text/json", "application/x-www-form-urlencoded" ],
        "produces": [ "application/json", "text/json" ],
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/cls_LotSearch" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": { "type": "object" }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "cls_CustomerContractWarehouse": {
      "type": "object",
      "properties": {
        "CustomerRateID": {
          "format": "int32",
          "type": "integer"
        },
        "Mode": { "type": "string" },
        "WareHouseID": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "cls_customerRateList": {
      "type": "object",
      "properties": {
        "CustomerRateListID": {
          "format": "int32",
          "type": "integer"
        },
        "CustomerID": {
          "format": "int32",
          "type": "integer"
        },
        "wefDate": {
          "format": "date-time",
          "type": "string"
        },
        "FromDate": {
          "format": "date-time",
          "type": "string"
        },
        "EndDate": {
          "format": "date-time",
          "type": "string"
        },
        "RateID": {
          "format": "int32",
          "type": "integer"
        },
        "BillingCycleID": {
          "format": "int32",
          "type": "integer"
        },
        "LTD_CustomerRateList": {
          "type": "array",
          "items": { "$ref": "#/definitions/TD_CustomerRateList" }
        },
        "LTD_CustomerStorageArea1": {
          "type": "array",
          "items": { "$ref": "#/definitions/TD_CustomerStorageArea1" }
        },
        "LTD_CustomerAddOnServices1": {
          "type": "array",
          "items": { "$ref": "#/definitions/TD_CustomerAddOnServices1" }
        },
        "LTD_WareHouses": {
          "type": "array",
          "items": { "$ref": "#/definitions/TD_WareHouses" }
        },
        "createdBy": {
          "format": "int32",
          "type": "integer"
        },
        "CompanyID": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "TD_CustomerRateList": {
      "type": "object",
      "properties": {
        "CustomerRateListID": {
          "format": "int32",
          "type": "integer"
        },
        "ProductID": {
          "format": "int32",
          "type": "integer"
        },
        "ServiceID": {
          "format": "int32",
          "type": "integer"
        },
        "ServiceName": { "type": "string" },
        "ChargedRate": {
          "format": "double",
          "type": "number"
        },
        "LabourRate": {
          "format": "double",
          "type": "number"
        }
      }
    },
    "TD_CustomerStorageArea1": {
      "type": "object",
      "properties": {
        "StorageAreaID": {
          "format": "int32",
          "type": "integer"
        },
        "SpaceAllocationPercentage": {
          "format": "int32",
          "type": "integer"
        },
        "StartDate": {
          "format": "date-time",
          "type": "string"
        },
        "EndDate": {
          "format": "date-time",
          "type": "string"
        },
        "Rate": {
          "format": "double",
          "type": "number"
        },
        "BillingCycleID": {
          "format": "int32",
          "type": "integer"
        },
        "IsUpdate": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "TD_CustomerAddOnServices1": {
      "type": "object",
      "properties": {
        "ServiceID": {
          "format": "int32",
          "type": "integer"
        },
        "StartDate": {
          "format": "date-time",
          "type": "string"
        },
        "EndDate": {
          "format": "date-time",
          "type": "string"
        },
        "Rate": {
          "format": "double",
          "type": "number"
        },
        "BillingCycleID": {
          "format": "int32",
          "type": "integer"
        },
        "IsUpdate": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "TD_WareHouses": {
      "type": "object",
      "properties": {
        "WarehouseID": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
...
    "cls_CustomerLimit": {
      "type": "object",
      "properties": {
        "CustomerLimitID": {
          "format": "int32",
          "type": "integer"
        },
        "CustomerID": {
          "format": "int32",
          "type": "integer"
        },
        "WarehouseID": {
          "format": "int32",
          "type": "integer"
        },
        "LimitAmount": {
          "format": "double",
          "type": "number"
        },
        "CreatedBy": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
...
    "cls_CustomersStock": {
      "type": "object",
      "properties": {
        "CustomerID": {
          "format": "int32",
          "type": "integer"
        },
        "WarehouseID": {
          "format": "int32",
          "type": "integer"
        },
        "LotNo": { "type": "string" }
      }
    },
...
    "cls_DockMaster": {
      "type": "object",
      "properties": {
        "dockID": {
          "format": "int32",
          "type": "integer"
        },
        "dockName": { "type": "string" },
        "dockNo": {
          "format": "int32",
          "type": "integer"
        },
        "isActive": { "type": "boolean" },
        "wareHouseID": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
...
    "cls_InwardUpdate": {
      "type": "object",
      "properties": {
        "inwardDID": {
          "format": "int32",
          "type": "integer"
        },
        "brandId": {
          "format": "int32",
          "type": "integer"
        },
        "itemsInPacket": { "type": "string" },
        "inQuantity": {
          "format": "double",
          "type": "number"
        },
        "lotno": { "type": "string" },
        "docID": { "type": "string" },
        "unLoadingBy": { "type": "string" },
        "remarks": { "type": "string" },
        "truckNo": { "type": "string" },
        "containerNo": { "type": "string" },
        "inwardID": {
          "format": "int32",
          "type": "integer"
        },
        "storageAreaID": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
...
    "cls_UserLogin": {
      "type": "object",
      "properties": {
        "UserName": { "type": "string" },
        "Password": { "type": "string" }
      }
    },
    "cls_LotSearch": {
      "type": "object",
      "properties": {
        "lotNo": { "type": "string" },
        "wareHouseID": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
...
  }
}